// src/app/api-client/model/fineTuneSeriesInfo.ts
/**
 * Horde Model Reference API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

/**
 * Information about a fine-tuning series.
 */
export interface FineTuneSeriesInfo {
  name: string;
  version?: string | null;
  author?: string | null;
  description?: string | null;
  homepage?: string | null;
}